move guard implementation into a module (WIP)
